Add loading indicator while saving
- Prevent user from performing actions that modify server multiple
  times while a request is pending.

+++ client/src/views/Game.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { Game, GameUtilities } from '../models/game.model';
 import { CreateGame } from './CreateGame';
 import { gameService } from '../services/GameService';
+import classNames from 'classnames';
 
 type GameProps = {
     game: Game;
@@ -9,16 +10,22 @@ type GameProps = {
 
 export const GameComponent: React.FC<GameProps> = props => {
     const [isEditing, setIsEditing] = useState<boolean>(false);
+    const [isSaving, setIsSaving] = useState<boolean>(false);
     if (isEditing) {
         const handleSubmit = async (
             event: React.FormEvent,
             game: Game
-        ): Promise<Game> => {
+        ): Promise<void> => {
             event.preventDefault();
-            const created = await gameService.updateGame(game);
+            if (isSaving) {
+                return;
+            }
+            setIsSaving(true);
+            await gameService.updateGame(game);
             // If it succeeds close the edit view
             setIsEditing(false);
-            return created;
+            setIsSaving(false);
+            return;
         };
 
         return (
@@ -27,13 +34,19 @@ export const GameComponent: React.FC<GameProps> = props => {
                 onSubmit={handleSubmit}
                 titleText="Update Game"
                 submitButtonText="Save"
+                loading={isSaving}
             />
         );
     }
-    const handleDelete = (e: React.FormEvent) => {
+    const handleDelete = async (e: React.FormEvent) => {
         e.preventDefault();
+        if (isSaving) {
+            return;
+        }
         if (window.confirm('Are you sure you want to delete this game?')) {
-            gameService.deleteGame(props.game.id);
+            setIsSaving(true);
+            await gameService.deleteGame(props.game.id);
+            setIsSaving(false);
         }
     };
     return (
@@ -41,11 +54,21 @@ export const GameComponent: React.FC<GameProps> = props => {
             <div className="game_title">
                 <span>Game: {props.game.title}</span>
                 <div className="game_title_controls">
-                    <button onClick={() => setIsEditing(true)}>
+                    <button
+                        className={classNames({
+                            loading: isSaving
+                        })}
+                        onClick={() => setIsEditing(true)}
+                    >
                         <span className="icon-pencil"></span>
                         Edit
                     </button>
-                    <button className="warning" onClick={e => handleDelete(e)}>
+                    <button
+                        className={classNames('warning', {
+                            loading: isSaving
+                        })}
+                        onClick={e => handleDelete(e)}
+                    >
                         <span className="icon-bin"></span>
                         Delete
                     </button>
